Add files via upload
Few updates to the code. Added another variable start to use on the open_price variable to iterate through the open price rows. Added an if else statement for the percent_change, since I need to take into account the event when open_price is 0, where in this case, the value for percent_change would be 0, taking into account all other scenarios with this code.